Refactor classes into their own namespaces
