+++ src/System.CommandLine.CommandHandler/CommandHandlerGenerator.cs
@@ -1,6 +1,4 @@
 ﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
@@ -13,8 +11,6 @@ namespace System.CommandLine.CommandHandler
 
         public void Execute(GeneratorExecutionContext context)
         {
-            //AddGeneratorClass(context);
-
             SyntaxReceiver rx = (SyntaxReceiver)context.SyntaxContextReceiver!;
 
             StringBuilder builder = new();
@@ -96,118 +92,6 @@ namespace System.CommandLine.Invocation
             context.AddSource("CommandHandlerGeneratorExtensions_Generated.g.cs", builder.ToString());
         }
 
-        private static void AddGeneratorClass(GeneratorExecutionContext context)
-        {
-            context.AddSource("GeneratedCommandHandler2.g.cs", @"
-// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System.CommandLine.Binding;
-using System.Reflection;
-using System.Threading.Tasks;
-
-namespace System.CommandLine.Invocation
-{
-    public static class CommandHandlerGeneratorExtensions_Generated
-    {
-private class GeneratedHandler_1 : ICommandHandler
-{
-    public GeneratedHandler_1(Action<string, IConsole, int> method,
-    Option<string> param1,
-    Option<int> param2)
-    {
-        Method = method;
-        Param1 = param1;
-        Param2 = param2;
-    }
-        
-    public Action<string, IConsole, int> Method { get; }
-    public Option<string> Param1 { get; }
-    public Option<int> Param2 { get; }
-        
-    public Task<int> InvokeAsync(InvocationContext context)
-    {
-        string value1 = context.ParseResult.ValueForOption(Param1);
-        int value2 = context.ParseResult.ValueForOption(Param2);
-        
-        Method.Invoke(value1, context.Console, value2);
-        
-        return Task.FromResult(0);
-    }
-}
-
-        private class GeneratedHandler_2 : ICommandHandler
-        {
-            public GeneratedHandler_2(Action<System.CommandLine.Tests.Invocation.CommandHandlerTests.Character, IConsole> method,
-            Option<string> param1,
-            Option<int> param2)
-            {
-                Method = method;
-                Param1 = param1;
-                Param2 = param2;
-            }
-        
-            public Action<System.CommandLine.Tests.Invocation.CommandHandlerTests.Character, IConsole> Method { get; }
-            public Option<string> Param1 { get; }
-            public Option<int> Param2 { get; }
-        
-            public Task<int> InvokeAsync(InvocationContext context)
-            {
-                string value1 = context.ParseResult.ValueForOption(Param1);
-                int value2 = context.ParseResult.ValueForOption(Param2);
-                System.CommandLine.Tests.Invocation.CommandHandlerTests.Character model = new(value1, value2);
-                Method.Invoke(model, context.Console);
-        
-                return Task.FromResult(0);
-            }
-        }
-
-        private class GeneratedHandler_3 : ICommandHandler
-        {
-            public GeneratedHandler_3(Action<System.CommandLine.Tests.Invocation.CommandHandlerTests.Character, IConsole> method,
-            Func<InvocationContext, System.CommandLine.Tests.Invocation.CommandHandlerTests.Character> modelBuilder)
-            {
-                Method = method;
-                ModelBuilder = modelBuilder;
-            }
-        
-            public Action<System.CommandLine.Tests.Invocation.CommandHandlerTests.Character, IConsole> Method { get; }
-            public Func<InvocationContext, System.CommandLine.Tests.Invocation.CommandHandlerTests.Character> ModelBuilder { get; }
-        
-            public Task<int> InvokeAsync(InvocationContext context)
-            {
-                System.CommandLine.Tests.Invocation.CommandHandlerTests.Character model = ModelBuilder(context);
-                Method.Invoke(model, context.Console);
-        
-                return Task.FromResult(0);
-            }
-        }
-
-        public static ICommandHandler Generate<TUnused>(this CommandHandlerGenerator handler, 
-            Action<string, IConsole, int> method,
-            Option<string> param1, Option<int> param2)
-        {
-            return new GeneratedHandler_1(method, param1, param2);
-        }
-
-        public static ICommandHandler Generate<TUnused>(this CommandHandlerGenerator handler, 
-            Action<System.CommandLine.Tests.Invocation.CommandHandlerTests.Character, IConsole> method,
-            Option<string> param1, Option<int> param2)
-        {
-            return new GeneratedHandler_2(method, param1, param2);
-        }
-
-        public static ICommandHandler Generate<TUnused, TUnused2>(this CommandHandlerGenerator handler, 
-            Action<System.CommandLine.Tests.Invocation.CommandHandlerTests.Character, IConsole> method,
-            Func<InvocationContext, System.CommandLine.Tests.Invocation.CommandHandlerTests.Character> modelBuilder)
-        {
-            return new GeneratedHandler_3(method, modelBuilder);
-        }
-    }
-}
-");
-        }
-
         public void Initialize(GeneratorInitializationContext context)
         {
 #if DEBUG
@@ -218,175 +102,6 @@ private class GeneratedHandler_1 : ICommandHandler
 #endif
 
             context.RegisterForSyntaxNotifications(() => new SyntaxReceiver());
-        }
-    }
-
-    public class SyntaxReceiver : ISyntaxContextReceiver
-    {
-        public List<DelegateInvocation> Invocations { get; } = new();
-
-        public void OnVisitSyntaxNode(GeneratorSyntaxContext context)
-        {
-            if (context.Node is InvocationExpressionSyntax invocationExpression &&
-                invocationExpression.Expression is MemberAccessExpressionSyntax memberAccess &&
-                memberAccess.Name.Identifier.Text == "Generate" &&
-                context.SemanticModel.GetSymbolInfo(invocationExpression) is { } invocationSymbol &&
-                invocationSymbol.Symbol is IMethodSymbol invokeMethodSymbol &&
-                invokeMethodSymbol.ReceiverType?.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat) == "global::System.CommandLine.Invocation.CommandHandlerGenerator" &&
-                (invokeMethodSymbol.TypeArguments.Length == 1 || invokeMethodSymbol.TypeArguments.Length == 2))
-            {
-                INamedTypeSymbol? iConsole = context.SemanticModel.Compilation.GetTypeByMetadataName("System.CommandLine.IConsole");
-                if (iConsole is null) return;
-
-                IReadOnlyList<ISymbol> delegateParameters = Array.Empty<ISymbol>();
-                //Check for model binding condition
-                if (invokeMethodSymbol.TypeArguments[0] is INamedTypeSymbol namedDelegateType &&
-                    namedDelegateType.TypeArguments.Length > 0)
-                {
-                    delegateParameters = namedDelegateType.TypeArguments.Cast<ISymbol>().ToList();
-                }
-
-                IReadOnlyList<ISymbol?> symbols = invocationExpression.ArgumentList.Arguments
-                    .Skip(1)
-                    .Select(x => context.SemanticModel.GetSymbolInfo(x.Expression).Symbol)
-                    .ToList();
-                if (symbols.Any(x => x is null)) return;
-                IReadOnlyList<Parameter> givenParameters = GetParameters(symbols!);
-                if (invokeMethodSymbol.TypeArguments.Length == 2)
-                {
-                    //System.Diagnostics.Debugger.Launch();
-                    var rawParameter = (RawParameter)givenParameters[0];
-                    var factoryParameter = new FactoryParameter(rawParameter, invokeMethodSymbol.Parameters[1].Type);
-                    givenParameters = new Parameter[] { factoryParameter }.Concat(givenParameters.Skip(1)).ToList();
-                }
-
-                SymbolEqualityComparer symbolEqualityComparer = SymbolEqualityComparer.Default;
-                HashSet<ISymbol> knownTypes = new(symbolEqualityComparer);
-                knownTypes.Add(iConsole);
-
-                if (IsMatch(delegateParameters, givenParameters, knownTypes))
-                {
-                    if (invokeMethodSymbol.TypeArguments.Length == 2)
-                    {
-                        var invocation = new FactoryModelBindingInvocation(invokeMethodSymbol.TypeArguments[0]);
-                        foreach (var parameter in PopulateParameters(delegateParameters, givenParameters, iConsole))
-                        {
-                            invocation.Parameters.Add(parameter);
-                        }
-                        Invocations.Add(invocation);
-                    }
-                    else
-                    {
-                        var invocation = new DelegateInvocation(invokeMethodSymbol.TypeArguments[0], 1);
-                        foreach (var parameter in PopulateParameters(delegateParameters, givenParameters, iConsole))
-                        {
-                            invocation.Parameters.Add(parameter);
-                        }
-                        Invocations.Add(invocation);
-                    }
-                }
-                else if (delegateParameters[0] is INamedTypeSymbol modelType)
-                {
-                    foreach (var ctor in modelType.Constructors.OrderByDescending(x => x.Parameters.Length))
-                    {
-                        var targetTypes =
-                            ctor.Parameters.Select(x => x.Type)
-                            .Concat(delegateParameters.Skip(1))
-                            .ToList();
-                        if (IsMatch(targetTypes, givenParameters, knownTypes))
-                        {
-                            var invocation = new ConstructorModelBindingInvocation(ctor, invokeMethodSymbol.TypeArguments[0]);
-                            foreach (var parameter in PopulateParameters(targetTypes, givenParameters, iConsole))
-                            {
-                                invocation.Parameters.Add(parameter);
-                            }
-                            Invocations.Add(invocation);
-                            break;
-                        }
-                    }
-                }
-
-                static bool IsMatch(
-                    IReadOnlyList<ISymbol> targetSymbols,
-                    IReadOnlyList<Parameter> providedSymbols,
-                    HashSet<ISymbol> knownTypes)
-                {
-                    SymbolEqualityComparer symbolEqualityComparer = SymbolEqualityComparer.Default;
-
-                    for (int i = 0, j = 0; i < targetSymbols.Count; i++)
-                    {
-                        if (j < providedSymbols.Count &&
-                            symbolEqualityComparer.Equals(providedSymbols[j].ValueType, targetSymbols[i]))
-                        {
-                            j++;
-                            //TODO: Handle the case where there are more provided symbols than needed
-                        }
-                        else if (!knownTypes.Contains(targetSymbols[i]))
-                        {
-                            return false;
-                        }
-                    }
-                    return true;
-                }
-            }
-        }
-
-        private static IReadOnlyList<Parameter> PopulateParameters(
-            IReadOnlyList<ISymbol> symbols,
-            IReadOnlyList<Parameter> givenParameters,
-            ITypeSymbol iConsole)
-        {
-            SymbolEqualityComparer symbolEqualityComparer = SymbolEqualityComparer.Default;
-            List<Parameter> parameters = new(givenParameters);
-            for (int i = 0; i < symbols.Count; i++)
-            {
-                if (symbolEqualityComparer.Equals(iConsole, symbols[i]))
-                {
-                    parameters.Insert(i, new ConsoleParameter(iConsole));
-                }
-            }
-            return parameters;
-        }
-
-        private static IReadOnlyList<Parameter> GetParameters(IEnumerable<ISymbol> symbols)
-        {
-            List<Parameter> parameters = new();
-            int parameterIndex = 1;
-            foreach (ISymbol symbol in symbols)
-            {
-                parameters.Add(GetParameter(symbol, $"Param{parameterIndex}"));
-                parameterIndex++;
-            }
-            return parameters;
-        }
-
-        private static Parameter GetParameter(ISymbol argumentSymbol, string localName)
-        {
-            return argumentSymbol switch
-            {
-                ILocalSymbol local => FromTypeSymbol(local.Type),
-                INamedTypeSymbol namedType => FromNamedTypeSymbol(namedType),
-                IMethodSymbol methodSymbol => FromTypeSymbol(methodSymbol.ReturnType),
-                _ => throw new NotImplementedException($"Cannot convert from '{argumentSymbol?.Kind}' {argumentSymbol?.ToDisplayString()}")
-            };
-
-            Parameter FromNamedTypeSymbol(INamedTypeSymbol namedTypeSymbol)
-            {
-                if (namedTypeSymbol.TypeArguments.Length > 0)
-                {
-                    return new Option(localName, namedTypeSymbol, namedTypeSymbol.TypeArguments[0]);
-                }
-                return new RawParameter(localName, namedTypeSymbol);
-            }
-
-            Parameter FromTypeSymbol(ITypeSymbol typeSymbol)
-            {
-                return typeSymbol switch
-                {
-                    INamedTypeSymbol namedType => FromNamedTypeSymbol(namedType),
-                    _ => throw new NotImplementedException($"Cannot convert from type symbol '{typeSymbol?.Kind}' {typeSymbol?.ToDisplayString()}")
-                };
-            }
         }
     }
 }
