+++ src/System.CommandLine.CommandHandler/Invocations/DelegateInvocation.cs
@@ -1,9 +1,10 @@
 ﻿using Microsoft.CodeAnalysis;
 using System.Collections.Generic;
+using System.CommandLine.CommandHandler.Parameters;
 using System.Linq;
 using System.Text;
 
-namespace System.CommandLine.CommandHandler
+namespace System.CommandLine.CommandHandler.Invocations
 {
     public class DelegateInvocation
     {
