+++ src/System.CommandLine.CommandHandler/Parameters/FactoryParameter.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.CodeAnalysis;
 
-namespace System.CommandLine.CommandHandler
+namespace System.CommandLine.CommandHandler.Parameters
 {
     public class FactoryParameter : Parameter
     {
